+++ core/node/node_test.go
@@ -1,4 +1,4 @@
-package util
+package node
 
 import (
 	"testing"
